@@ -1,13 +1,13 @@
 source 'https://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.2.6'
+gem 'rails', '7.0.3'
 
-gem 'sass-rails', '~> 5.0'
-#gem 'coffee-rails', '~> 4.1.0'
-#gem 'uglifier', '>= 4.2.0'
-#terser for minifying js es6 (not done correctly by uglifier)
-gem 'terser', '~> 1.1', '>= 1.1.5'
+gem 'jsbundling-rails'
+
+gem 'sassc-rails' #sass-rails replacent
+gem 'terser' #ugilifer replacent
+
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
 # gem 'duktape'
@@ -18,14 +18,35 @@ gem 'jquery-rails'
 gem 'jquery-ui-rails'
 gem 'select2-rails'
 
+
+# The original asset pipeline for Rails [https://github.com/rails/sprockets-rails]
+gem 'sprockets-rails'
+
+
+# Use the Puma web server [https://github.com/puma/puma]
+gem 'puma', '~> 5.0'
+
+# Use JavaScript with ESM import maps [https://github.com/rails/importmap-rails]
+gem 'importmap-rails'
+
+# Hotwire's SPA-like page accelerator [https://turbo.hotwired.dev]
+gem 'turbo-rails'
+
+# Hotwire's modest JavaScript framework [https://stimulus.hotwired.dev]
+gem 'stimulus-rails'
+
+# Build JSON APIs with ease [https://github.com/rails/jbuilder]
+#gem "jbuilder"
+
+# Windows does not include zoneinfo files, so bundle the tzinfo-data gem
+gem 'tzinfo-data', platforms: %i[ mingw mswin x64_mingw jruby ]
+
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
 
-# To use Jbuilder templates for JSON
-# gem 'jbuilder'
+# Reduces boot times through caching; required in config/boot.rb
 
-# To use debugger
-# gem 'debugger'
+gem 'bootsnap', require: false
 
 gem 'cube-ruby', require: 'cube'
 gem 'dalli'
@@ -35,21 +56,22 @@ gem 'i18n'
 gem 'iconv'
 gem 'multi_json'
 gem 'mysql2', '0.5.2'
-gem 'nokogiri'
 gem 'oj'
 gem 'open_uri_redirections'
 gem 'pry'
+gem 'psych', '< 4'
 gem 'rack-mini-profiler'
 gem 'rails_autolink'
 gem 'rdoc'
-gem 'recaptcha', '~> 5.2'
+gem 'recaptcha', '~> 5.9.0'
 gem 'rest-client'
 gem 'stackprof', require: false
 gem 'thin'
+gem 'view_component', '~> 2.72'
 gem 'turnout'
 gem 'will_paginate', '~> 3.0'
 
-gem 'ontologies_api_client', git: "https://github.com/ontoportal-lirmm/ontologies_api_ruby_client.git", tag: "v2.0.2-merged"
+gem 'ontologies_api_client', git: 'https://github.com/ontoportal-lirmm/ontologies_api_ruby_client.git', branch: 'master'
 
 
 group :staging, :production do
@@ -73,6 +95,11 @@ group :development do
   gem 'brakeman', require: false
   gem 'rubocop', require: false
   # gem 'i18n-debug'
+  # See https://guides.rubyonrails.org/debugging_rails_applications.html#debugging-with-the-debug-gem
+  gem 'debug', platforms: %i[ mri mingw x64_mingw ]
+
+  # Use console on exceptions pages [https://github.com/rails/web-console]
+  gem 'web-console'
 end
 
 group :test, :development do
@@ -80,5 +107,9 @@ group :test, :development do
 end
 
 group :test do
+  # Use system testing [https://guides.rubyonrails.org/testing.html#system-testing]
   gem 'capybara'
+  gem 'selenium-webdriver'
+  gem 'webdrivers'
 end
+
