@@ -1,20 +1,21 @@
 require 'cgi'
+
 class MappingsController < ApplicationController
   include ActionView::Helpers::NumberHelper
-
+  include TurboHelper
+  include MappingStatistics
+  include MappingsHelper
+  include TurboHelper
   layout :determine_layout
-  before_action :authorize_and_redirect, :only=>[:create,:new,:destroy]
+  before_action :authorize_and_redirect, only: [:create, :new, :destroy]
 
-  MAPPINGS_URL = "#{LinkedData::Client.settings.rest_url}/mappings"
-  EXTERNAL_MAPPINGS_GRAPH = "http://data.bioontology.org/metadata/ExternalMappings"
-  INTERPORTAL_MAPPINGS_GRAPH = "http://data.bioontology.org/metadata/InterportalMappings"
   EXTERNAL_URL_PARAM_STR = "mappings:external"
   INTERPORTAL_URL_PARAM_STR = "interportal:"
 
   INTERPORTAL_HASH = $INTERPORTAL_HASH ||= {}
 
   def index
-    ontology_list = LinkedData::Client::Models::Ontology.all.select {|o| !o.summaryOnly}
+    ontology_list = LinkedData::Client::Models::Ontology.all.select { |o| !o.summaryOnly }
     ontologies_mapping_count = LinkedData::Client::HTTP.get("#{MAPPINGS_URL}/statistics/ontologies")
     ontologies_hash = {}
     ontology_list.each do |ontology|
@@ -27,25 +28,23 @@ class MappingsController < ApplicationController
     # end
 
     @options = {}
-    if ontologies_mapping_count
-      ontologies_mapping_count.members.each do |ontology_acronym|
-        # Adding external and interportal mappings to the dropdown list
-        if ontology_acronym.to_s == EXTERNAL_MAPPINGS_GRAPH
-          mapping_count = ontologies_mapping_count[ontology_acronym.to_s] || 0
-          select_text = "External Mappings (#{number_with_delimiter(mapping_count, delimiter: ',')})" if mapping_count > 0
-          ontology_acronym = EXTERNAL_URL_PARAM_STR
-        elsif ontology_acronym.to_s.start_with?(INTERPORTAL_MAPPINGS_GRAPH)
-          mapping_count = ontologies_mapping_count[ontology_acronym.to_s] || 0
-          select_text = "Interportal Mappings - #{ontology_acronym.to_s.split("/")[-1].upcase} (#{number_with_delimiter(mapping_count, delimiter: ',')})" if mapping_count > 0
-          ontology_acronym = INTERPORTAL_URL_PARAM_STR + ontology_acronym.to_s.split("/")[-1]
-        else
-          ontology = ontologies_hash[ontology_acronym.to_s]
-          mapping_count = ontologies_mapping_count[ontology_acronym] || 0
-          next unless ontology && mapping_count > 0
-          select_text = "#{ontology.name} - #{ontology.acronym} (#{number_with_delimiter(mapping_count, delimiter: ',')})"
-        end
-        @options[select_text] = ontology_acronym
+    ontologies_mapping_count&.members&.each do |ontology_acronym|
+      # Adding external and interportal mappings to the dropdown list
+      if ontology_acronym.to_s == EXTERNAL_MAPPINGS_GRAPH
+        mapping_count = ontologies_mapping_count[ontology_acronym.to_s] || 0
+        select_text = "External Mappings (#{number_with_delimiter(mapping_count, delimiter: ',')})" if mapping_count >= 0
+        ontology_acronym = EXTERNAL_URL_PARAM_STR
+      elsif ontology_acronym.to_s.start_with?(INTERPORTAL_MAPPINGS_GRAPH)
+        mapping_count = ontologies_mapping_count[ontology_acronym.to_s] || 0
+        select_text = "Interportal Mappings - #{ontology_acronym.to_s.split("/")[-1].upcase} (#{number_with_delimiter(mapping_count, delimiter: ',')})" if mapping_count >= 0
+        ontology_acronym = INTERPORTAL_URL_PARAM_STR + ontology_acronym.to_s.split("/")[-1]
+      else
+        ontology = ontologies_hash[ontology_acronym.to_s]
+        mapping_count = ontologies_mapping_count[ontology_acronym] || 0
+        next unless ontology && mapping_count > 0
+        select_text = "#{ontology.name} - #{ontology.acronym} (#{number_with_delimiter(mapping_count, delimiter: ',')})"
       end
+      @options[select_text] = ontology_acronym
     end
 
     @options = @options.sort
@@ -53,45 +52,57 @@ class MappingsController < ApplicationController
 
   def count
     @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:id]).first
-    if @ontology
-      @ontology_id = @ontology.acronym
-      @ontology_label = @ontology.name
-      counts = LinkedData::Client::HTTP.get("#{MAPPINGS_URL}/statistics/ontologies/#{params[:id]}")
-    else
-      @ontology = params[:id]
-      @ontology_id = @ontology
-      @ontology_label = params[:id].split(":")[-1]
-      if @ontology_label == "external"
-        counts = LinkedData::Client::HTTP.get("#{MAPPINGS_URL}/statistics/external")
-      elsif params[:id].split(":")[0] == "interportal"
-        counts = LinkedData::Client::HTTP.get("#{MAPPINGS_URL}/statistics/interportal/#{@ontology_label}")
-      end
-    end
+    @ontology_acronym = @ontology&.acronym || params[:id]
+    @mapping_counts = mapping_counts(@ontology_acronym)
+    render partial: 'count'
+  end
 
-    @ontologies_mapping_count = []
-    counts.members.each do |acronym|
-      count = counts[acronym]
-      ontology = LinkedData::Client::Models::Ontology.find_by_acronym(acronym.to_s).first
-      if ontology
-        onto_info = {:id => ontology.id, :name => ontology.name, :viewOf => ontology.viewOf}
-      else
-        if acronym.to_s == EXTERNAL_MAPPINGS_GRAPH
-          onto_info = {:id => acronym.to_s, :name => "External Mappings", :viewOf => nil}
-          @ontologies_mapping_count << {'ontology' => onto_info, 'count' => count}
-        elsif acronym.to_s.start_with?(INTERPORTAL_MAPPINGS_GRAPH)
-          onto_info = {:id => acronym.to_s, :name => "#{acronym.to_s.split("/")[-1].upcase} Interportal", :viewOf => nil}
-          @ontologies_mapping_count << {'ontology' => onto_info, 'count' => count}
-        end
-      end
-      next unless ontology
-      @ontologies_mapping_count << {'ontology' => onto_info, 'count' => count}
-    end
-    @ontologies_mapping_count.sort! {|a,b| a['ontology'][:name].downcase <=> b['ontology'][:name].downcase } unless @ontologies_mapping_count.nil? || @ontologies_mapping_count.length == 0
+  def loader
+    @example_code = [{
+                       "classes": ["http://bioontology.org/ontologies/BiomedicalResourceOntology.owl#Image_Algorithm",
+                                   "http://purl.org/incf/ontology/Computational_Neurosciences/cno_alpha.owl#cno_0000202"],
 
-    render :partial => 'count'
+                       "name": 'This is the mappings produced to test the bulk load',
+                       "source": 'https://w3id.org/semapv/LexicalMatching',
+                       "comment": 'mock data',
+                       "relation": [
+                         'http://www.w3.org/2002/07/owl#subClassOf'
+                       ],
+                       "subject_source_id": 'http://bioontology.org/ontologies/BiomedicalResources.owl',
+                       "object_source_id": 'http://purl.org/incf/ontology/Computational_Neurosciences/cno_alpha.owl',
+                       "source_name": 'https://w3id.org/sssom/mapping/tests/data/basic.tsv',
+                       "source_contact_info": 'orcid:1234,orcid:5678',
+                       "date": '2020-05-30'
+                     }]
+    render partial: 'mappings/bulk_loader/loader'
+  end
+
+  def loader_process
+    response = LinkedData::Client::HTTP.post('/mappings/load', file: params[:file])
+    errors = response.errors
+    errors = errors.to_h.except(:links, :context) if errors
+
+    created = response.created
+    respond_to do |format|
+      format.turbo_stream do
+        render turbo_stream: turbo_stream.replace('file_loader_result',
+                                                  partial: 'mappings/bulk_loader/loaded_mappings',
+                                                  locals: { errors: errors, created: created })
+      end
+      format.html { redirect_to mappings_path }
+    end
   end
 
   def show
+    @mapping = request_mapping
+    mapping_form(mapping: @mapping)
+    respond_to do |format|
+      format.html { render 'mappings/edit', layout: false }
+    end
+
+  end
+
+  def show_mappings
     page = params[:page] || 1
     @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:id]).first
     @target_ontology = LinkedData::Client::Models::Ontology.find(params[:target])
@@ -123,7 +134,7 @@ class MappingsController < ApplicationController
 
     ontologies = [ontology_acronym, target_acronym]
 
-    @mapping_pages = LinkedData::Client::HTTP.get(MAPPINGS_URL, {page: page, ontologies: ontologies.join(",")})
+    @mapping_pages = LinkedData::Client::HTTP.get("#{MAPPINGS_URL}", { page: page, ontologies: ontologies.join(',') })
     @mappings = @mapping_pages.collection
     @delete_mapping_permission = check_delete_mapping_permission(@mappings)
 
@@ -138,131 +149,198 @@ class MappingsController < ApplicationController
 
     # This converts the mappings into an object that can be used with the pagination plugin
     @page_results = WillPaginate::Collection.create(@mapping_pages.page, @mapping_pages.collection.length, total_results) do |pager|
-       pager.replace(@mapping_pages.collection)
+      pager.replace(@mapping_pages.collection)
     end
 
-    render :partial => 'show'
+    render partial: 'show'
   end
 
   def get_concept_table
     @ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:ontologyid]).first
-    @concept = @ontology.explore.single_class({full: true}, params[:conceptid])
+    @concept = @ontology.explore.single_class({ full: true }, params[:conceptid])
 
     @mappings = @concept.explore.mappings
 
     @delete_mapping_permission = check_delete_mapping_permission(@mappings)
-
-    render :partial => "mapping_table"
+    render partial: 'mappings/concept_mappings'
   end
 
   def new
-    @ontology_from = LinkedData::Client::Models::Ontology.find(params[:ontology_from])
-    @ontology_to = LinkedData::Client::Models::Ontology.find(params[:ontology_to])
-    @concept_from = @ontology_from.explore.single_class({full: true}, params[:conceptid_from]) if @ontology_from
-    @concept_to = @ontology_to.explore.single_class({full: true}, params[:conceptid_to]) if @ontology_to
-
-    # Defaults just in case nothing gets provided
-    @ontology_from ||= LinkedData::Client::Models::Ontology.new
-    @ontology_to ||= LinkedData::Client::Models::Ontology.new
-    @concept_from ||= LinkedData::Client::Models::Class.new
-    @concept_to ||= LinkedData::Client::Models::Class.new
-
-    @mapping_relation_options = [
-      ["Identical (skos:exactMatch)", "http://www.w3.org/2004/02/skos/core#exactMatch"],
-      ["Similar (skos:closeMatch)",   "http://www.w3.org/2004/02/skos/core#closeMatch"],
-      ["Related (skos:relatedMatch)", "http://www.w3.org/2004/02/skos/core#relatedMatch"],
-      ["Broader (skos:broadMatch)",   "http://www.w3.org/2004/02/skos/core#broadMatch"],
-      ["Narrower (skos:narrowMatch)", "http://www.w3.org/2004/02/skos/core#narrowMatch"],
-      ["Translation (gold:translation)", "http://purl.org/linguistics/gold/translation"],
-      ["Free Translation (gold:freeTranslation)", "http://purl.org/linguistics/gold/freeTranslation"]
-    ]
-
+    mapping_form
     respond_to do |format|
-      format.js
-    end    
-  end
-
-  def new_external
-    @ontology_from = LinkedData::Client::Models::Ontology.find(params[:ontology_from])
-    @ontology_to = LinkedData::Client::Models::Ontology.find(params[:ontology_to])
-    @concept_from = @ontology_from.explore.single_class({full: true}, params[:conceptid_from]) if @ontology_from
-    @concept_to = @ontology_to.explore.single_class({full: true}, params[:conceptid_to]) if @ontology_to
-
-    @interportal_options = []
-    INTERPORTAL_HASH.each do |key, value|
-      @interportal_options.push([key, key])
-    end
-
-    # Defaults just in case nothing gets provided
-    @ontology_from ||= LinkedData::Client::Models::Ontology.new
-    @ontology_to ||= LinkedData::Client::Models::Ontology.new
-    @concept_from ||= LinkedData::Client::Models::Class.new
-    @concept_to ||= LinkedData::Client::Models::Class.new
-
-    @mapping_relation_options = [
-      ["Identical (skos:exactMatch)", "http://www.w3.org/2004/02/skos/core#exactMatch"],
-      ["Similar (skos:closeMatch)",   "http://www.w3.org/2004/02/skos/core#closeMatch"],
-      ["Related (skos:relatedMatch)", "http://www.w3.org/2004/02/skos/core#relatedMatch"],
-      ["Broader (skos:broadMatch)",   "http://www.w3.org/2004/02/skos/core#broadMatch"],
-      ["Narrower (skos:narrowMatch)", "http://www.w3.org/2004/02/skos/core#narrowMatch"],
-      ["Translation (gold:translation)", "http://purl.org/linguistics/gold/translation"],
-      ["Free Translation (gold:freeTranslation)", "http://purl.org/linguistics/gold/freeTranslation"]
-    ]
-
-    respond_to do |format|
-      format.js
+      format.html { render action: 'new', layout: false }
     end
   end
 
   # POST /mappings
   # POST /mappings.xml
   def create
-    if params.has_key?(:mapping_type)
-      # Means its an external or interportal mapping
-      target_ontology = params[:map_to_bioportal_ontology_id]
-      target = params[:map_to_bioportal_full_id]
-    else
-      # Means it's a regular internal mapping
-      target_ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:map_to_bioportal_ontology_id]).first
-      target = target_ontology.explore.single_class(params[:map_to_bioportal_full_id]).id
-      target_ontology = target_ontology.id
+    values, @concept = mapping_form_values
+    errors = valid_values?(values)
+    if errors.empty?
+      @mapping, = LinkedData::Client::Models::Mapping.new(values: values)
+      @mapping_saved = @mapping.save
+      errors << @mapping_saved.errors if @mapping_saved.errors
     end
-    source_ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:map_from_bioportal_ontology_id]).first
-    source = source_ontology.explore.single_class(params[:map_from_bioportal_full_id])
-    values = {
-      classes: {
-        source.id => source_ontology.id,
-        target => target_ontology
-      },
-      creator: session[:user].id,
-      relation: params[:mapping_relation],
-      comment: params[:mapping_comment]
-    }
-    @mapping = LinkedData::Client::Models::Mapping.new(values: values)
-    @mapping_saved = @mapping.save
-    if @mapping_saved.errors
-      render text: @mapping_saved.errors[0], status: :bad_request
-    else
-      @delete_mapping_permission = check_delete_mapping_permission(@mapping_saved)
-      render :json => @mapping_saved
+
+    respond_to do |format|
+      format.turbo_stream do
+        if !errors.empty?
+          render turbo_stream: alert_error { JSON.pretty_generate errors }
+        else
+          @delete_mapping_permission = check_delete_mapping_permission(@mapping_saved)
+          mapping = LinkedData::Client::Models::Mapping.find(@mapping_saved.id)
+          render turbo_stream: [
+            alert(type: 'success') { 'Mapping created' },
+            prepend('concept_mappings_table_content', partial: 'show_line', locals: { map: mapping, concept: @concept })
+          ]
+        end
+      end
     end
+  end
+
+  def update
+    values, @concept = mapping_form_values
+    @mapping = request_mapping
+    errors = valid_values?(values)
+    if errors.empty?
+      map_uri = "#{MAPPINGS_URL}/#{@mapping.id.split('/').last}"
+      response = LinkedData::Client::HTTP.patch(map_uri, values)
+      errors <<  response.body if response.status != 204
+    end
+
+    respond_to do |format|
+      format.turbo_stream do
+        if !errors.empty? || @mapping.nil?
+          render_turbo_stream(alert_error { JSON.pretty_generate errors })
+        else
+          render_turbo_stream(
+            alert_success { 'Mapping updated' },
+            replace(@mapping.id.split('/').last, partial: 'show_line', locals: { map: request_mapping, concept: @concept })
+          )
+        end
+      end
+    end
+
   end
 
   def destroy
-    errors = []
-    successes = []
-    mapping_ids = params[:mappingids].split(",")
-    mapping_ids.each do |map_id|
-      begin
-        map_uri = "#{MAPPINGS_URL}/#{CGI.escape(map_id)}"
-        result = LinkedData::Client::HTTP.delete(map_uri)
-        raise Exception if !result.nil? #&& result["errorCode"]
-        successes << map_id
-      rescue Exception => e
-        errors << map_id
-      end
+    error = nil
+    success_text = ''
+    map_id = params[:id].gsub(':/', '://')
+    map_uri = "#{MAPPINGS_URL}/#{CGI.escape(map_id)}"
+    result = LinkedData::Client::HTTP.delete(map_uri)
+    if result.status == 204
+      success_text = "#{map_id} deleted successfully"
+    else
+      error = result.body
     end
-    render :json => { :success => successes, :error => errors }
+    respond_to do |format|
+      format.turbo_stream do
+        if error.nil?
+          render turbo_stream: [
+            alert(type: 'success') { success_text },
+            turbo_stream.remove(map_id.split('/').last)
+          ]
+
+        else
+          render alert(type: 'danger') { error }
+        end
+      end
+      format.html { render json: { success: success_text, error: error } }
+    end
+
   end
 
+  private
+
+  def mapping_form(mapping: nil)
+    if mapping
+      mapping.classes.each do |cls|
+        if cls.id.eql?(params[:conceptid_from])
+          @concept_from = cls
+          @ontology_from = cls.explore.ontology
+        else
+          @concept_to = cls
+          @mapping_type = if inter_portal_mapping?(@concept_to)
+                            'interportal'
+                          elsif internal_mapping?(@concept_to)
+                            'internal'
+                          else
+                            'external'
+                          end
+          set_mapping_target(concept_to_id: @concept_to.id, ontology_to: @concept_to.links['ontology'],
+                             mapping_type: @mapping_type)
+        end
+      end
+    else
+      mapping = LinkedData::Client::Models::Mapping.new
+      @ontology_from = LinkedData::Client::Models::Ontology.find(params[:ontology_from])
+      @ontology_to = LinkedData::Client::Models::Ontology.find(params[:ontology_to])
+      @concept_from = @ontology_from.explore.single_class({ full: true }, params[:conceptid_from]) if @ontology_from
+      if @ontology_to
+       @concept_to = @ontology_to.explore.single_class({ full: true }, params[:conceptid_to])
+       @map_to_bioportal_ontology_id = @ontology_to.id
+      end
+    end
+
+    @interportal_options = []
+    INTERPORTAL_HASH.each do |key, value|
+      @interportal_options.push([key, value['api']])
+    end
+
+
+    @mapping_relation_options = [
+      ["Identical (skos:exactMatch)", "http://www.w3.org/2004/02/skos/core#exactMatch"],
+      ["Similar (skos:closeMatch)", "http://www.w3.org/2004/02/skos/core#closeMatch"],
+      ["Related (skos:relatedMatch)", "http://www.w3.org/2004/02/skos/core#relatedMatch"],
+      ["Broader (skos:broadMatch)", "http://www.w3.org/2004/02/skos/core#broadMatch"],
+      ["Narrower (skos:narrowMatch)", "http://www.w3.org/2004/02/skos/core#narrowMatch"],
+      ["Translation (gold:translation)", "http://purl.org/linguistics/gold/translation"],
+      ["Free Translation (gold:freeTranslation)", "http://purl.org/linguistics/gold/freeTranslation"]
+    ]
+    @mapping_name = mapping.process&.name
+    @mapping_comment = mapping.process&.comment
+    @mapping_source_name = mapping.process&.source_name
+    @mapping_source_contact_info = mapping.process&.source_contact_info
+    @mapping_source = mapping.process&.source
+    @selected_relation = mapping.process.nil? ? @mapping_relation_options.first : mapping.process.relation.first
+  end
+
+  def mapping_form_values
+    target_ontology, target, external_mapping = get_mappings_target
+    source_ontology = LinkedData::Client::Models::Ontology.find_by_acronym(params[:map_from_bioportal_ontology_id]).first
+    source = source_ontology.explore.single_class(params[:map_from_bioportal_full_id])
+    values = {
+      classes: [
+        source.id,
+        target
+      ],
+      subject_source_id: source_ontology.id,
+      object_source_id: target_ontology,
+      creator: session[:user].id,
+      external_mapping: external_mapping,
+      relation: Array(params[:mapping][:relation]),
+      source_contact_info: params[:mapping][:source_contact_info],
+      source_name: params[:mapping][:source_name],
+      name: params[:mapping][:name],
+      comment: params[:mapping][:comment]
+    }
+
+    [values, source]
+  end
+
+  def request_mapping
+    mapping = LinkedData::Client::Models::Mapping.find(params[:id])
+    not_found if mapping.nil? || mapping.errors
+    mapping
+  end
+
+  def valid_values?(values)
+    errors = []
+    if values[:classes].reject(&:blank?).size != 2
+      errors << 'Source and target concepts need to be specified'
+    end
+    errors
+  end
 end
