@@ -20,11 +20,11 @@ module OntologiesHelper
   def display_data_catalog(sub)
     if !sub.send("includedInDataCatalog").nil? && sub.send("includedInDataCatalog").any?
       # Buttons for data catalogs
-      return content_tag(:section, {:class => "ont-metadata-card ont-included-in-data-catalog-card"}) do
-        concat(content_tag(:div, {:class => "ont-section-toolbar"}) do
-          concat(content_tag(:header, "includedInDataCatalog", {:class => "pb-2 font-weight-bold"}))
+      return content_tag(:section, { :class => "ont-metadata-card ont-included-in-data-catalog-card" }) do
+        concat(content_tag(:div, { :class => "ont-section-toolbar" }) do
+          concat(content_tag(:header, "includedInDataCatalog", { :class => "pb-2 font-weight-bold" }))
         end)
-        concat(content_tag(:div, {:class => ""}) do
+        concat(content_tag(:div, { :class => "" }) do
           sub.send("includedInDataCatalog").each do |catalog|
             catalog_btn_label = catalog
             $DATA_CATALOG_VALUES.each do |cat_uri, cat_label|
@@ -33,7 +33,7 @@ module OntologiesHelper
                 break;
               end
             end
-            concat(content_tag(:a, catalog_btn_label, {:class => "btn btn-primary", :href => catalog, :target => "_blank"}))
+            concat(content_tag(:a, catalog_btn_label, { :class => "btn btn-primary", :href => catalog, :target => "_blank" }))
           end
         end)
       end
@@ -49,16 +49,16 @@ module OntologiesHelper
     logo_attributes.each do |metadata|
       if !sub.send(metadata).nil?
         puts sub.send(metadata)
-        logo_html.concat(content_tag(:section, {:class => "ont-metadata-card ont-logo-depiction-card"}) do
-          concat(content_tag(:div, {:class => "ont-section-toolbar"}) do
-            concat(content_tag(:header, metadata.capitalize, {:class => "pb-2 font-weight-bold"}))
+        logo_html.concat(content_tag(:section, { :class => "ont-metadata-card ont-logo-depiction-card" }) do
+          concat(content_tag(:div, { :class => "ont-section-toolbar" }) do
+            concat(content_tag(:header, metadata.capitalize, { :class => "pb-2 font-weight-bold" }))
           end)
-          concat(content_tag(:div, {:class => ""}) do
-            concat(content_tag(:a, {:href => sub.send(metadata), :title => sub.send(metadata),
-                             :target => "_blank", :style=>"border-width:0;"}) do
+          concat(content_tag(:div, { :class => "" }) do
+            concat(content_tag(:a, { :href => sub.send(metadata), :title => sub.send(metadata),
+                                     :target => "_blank", :style => "border-width:0;" }) do
 
-              concat(content_tag(:img, "",{:title => sub.send(metadata),
-                                           :style=>"border-width:0;max-width: 100%;", :src=>sub.send(metadata).to_s}))
+              concat(content_tag(:img, "", { :title => sub.send(metadata),
+                                             :style => "border-width:0;max-width: 100%;", :src => sub.send(metadata).to_s }))
             end)
           end)
         end)
@@ -114,10 +114,10 @@ module OntologiesHelper
                   concat(content_tag(:td, "Natural Language", " "))
                   # Display naturalLanguage as flag
                   concat(content_tag(:td) do
-                    concat(content_tag(:ul, {:class => "f32"}) do
+                    concat(content_tag(:ul, { :class => "f32" }) do
                       lang_codes.each do |lang_code|
                         if lang_code.length == 2
-                          concat(content_tag(:li, "", {:class => "flag #{lang_code}", :style => "margin-right: 0.5em;"}))
+                          concat(content_tag(:li, "", { :class => "flag #{lang_code}", :style => "margin-right: 0.5em;" }))
                         else
                           concat(content_tag(:li, lang_code))
                         end
@@ -170,36 +170,36 @@ module OntologiesHelper
                 if (metadata.to_s.eql?("hasLicense"))
                   if (sub.send(metadata).to_s.start_with?("http://creativecommons.org/licenses") || sub.send(metadata).start_with?("https://creativecommons.org/licenses"))
                     concat(content_tag(:td) do
-                      concat(content_tag(:a, {:rel => "license", :alt=>"Creative Commons License",
-                                              :href => sub.send(metadata), :target => "_blank", :style=>"border-width:0", :title => sub.send(metadata),
-                                              :src=>"https://i.creativecommons.org/l/by/4.0/88x31.png"}) do
+                      concat(content_tag(:a, { :rel => "license", :alt => "Creative Commons License",
+                                               :href => sub.send(metadata), :target => "_blank", :style => "border-width:0", :title => sub.send(metadata),
+                                               :src => "https://i.creativecommons.org/l/by/4.0/88x31.png" }) do
 
-                        concat(content_tag(:img, "",{:rel => "license", :alt=>"Creative Commons License", :title => sub.send(metadata),
-                                                     :style=>"border-width:0", :src=>"https://i.creativecommons.org/l/by/4.0/88x31.png"}))
+                        concat(content_tag(:img, "", { :rel => "license", :alt => "Creative Commons License", :title => sub.send(metadata),
+                                                       :style => "border-width:0", :src => "https://i.creativecommons.org/l/by/4.0/88x31.png" }))
                       end)
                     end)
 
                   elsif (sub.send(metadata).to_s.start_with?("http://opensource.org/licenses") || sub.send(metadata).start_with?("https://opensource.org/licenses"))
                     concat(content_tag(:td) do
-                      concat(content_tag(:a, {:rel => "license", :alt=>"Open Source License",
-                                              :href => sub.send(metadata), :title => sub.send(metadata),:target => "_blank", :style=>"border-width:0;",
-                                              :src=>"https://opensource.org/files/osi_logo_bold_100X133_90ppi.png"}) do
+                      concat(content_tag(:a, { :rel => "license", :alt => "Open Source License",
+                                               :href => sub.send(metadata), :title => sub.send(metadata), :target => "_blank", :style => "border-width:0;",
+                                               :src => "https://opensource.org/files/osi_logo_bold_100X133_90ppi.png" }) do
 
-                        concat(content_tag(:img, "",{:rel => "license", :alt=>"Open Source License", :title => sub.send(metadata),
-                                                     :style=>"height: 80px; border-width:0;", :src=>"https://opensource.org/files/osi_logo_bold_100X133_90ppi.png"}))
+                        concat(content_tag(:img, "", { :rel => "license", :alt => "Open Source License", :title => sub.send(metadata),
+                                                       :style => "height: 80px; border-width:0;", :src => "https://opensource.org/files/osi_logo_bold_100X133_90ppi.png" }))
                       end)
                     end)
 
                   else
                     concat(content_tag(:td) do
-                      concat(content_tag(:a, sub.send(metadata), {:rel => "license", :href => sub.send(metadata), :target => "_blank"}))
+                      concat(content_tag(:a, sub.send(metadata), { :rel => "license", :href => sub.send(metadata), :target => "_blank" }))
                     end)
                   end
 
                 elsif (metadata.to_s.eql?("endpoint") && (sub.send(metadata).start_with?("http://sparql.") || sub.send(metadata).start_with?("https://sparql.")))
                   concat(content_tag(:td) do
-                    concat(content_tag(:a, {:href => sub.send(metadata), :title => sub.send(metadata),
-                                            :target => "_blank", :style=>"border-width:0;"}) do
+                    concat(content_tag(:a, { :href => sub.send(metadata), :title => sub.send(metadata),
+                                             :target => "_blank", :style => "border-width:0;" }) do
 
                       concat(image_tag('logos/sparql_logo.png', title: sub.send(metadata), class: 'logo'))
                     end)
@@ -211,13 +211,13 @@ module OntologiesHelper
                   if sub.send(metadata).to_s.split("/").length < 6
                     # for ontologies/ACRONYM we redirect to the UI url
                     concat(content_tag(:td) do
-                      concat(content_tag(:a, sub.send(metadata).to_s.split("/")[4..-1].join("/"), {:class=>"btn btn-primary",
-                                                                                                   :href => sub.send(metadata).sub("data.", ""), :target => "_blank", :title => sub.send(metadata)}))
+                      concat(content_tag(:a, sub.send(metadata).to_s.split("/")[4..-1].join("/"), { :class => "btn btn-primary",
+                                                                                                    :href => sub.send(metadata).sub("data.", ""), :target => "_blank", :title => sub.send(metadata) }))
                     end)
                   else
                     concat(content_tag(:td) do
-                      concat(content_tag(:a, sub.send(metadata).to_s.split("/")[4..-1].join("/"), {:class=>"btn btn-primary",
-                                                                                                   :href => sub.send(metadata), :target => "_blank", :title => sub.send(metadata)}))
+                      concat(content_tag(:a, sub.send(metadata).to_s.split("/")[4..-1].join("/"), { :class => "btn btn-primary",
+                                                                                                    :href => sub.send(metadata), :target => "_blank", :title => sub.send(metadata) }))
                     end)
                   end
 
@@ -242,7 +242,7 @@ module OntologiesHelper
     html.join("")
   end
 
-  def count_links(ont_acronym, page_name='summary', count=0)
+  def count_links(ont_acronym, page_name = 'summary', count = 0)
     ont_url = "/ontologies/#{ont_acronym}"
     if count.nil? || count == 0
       return "0"
@@ -270,7 +270,7 @@ module OntologiesHelper
     else
       uri = submission.id + "/download?apikey=#{get_apikey}"
       link = "<a href='#{uri}' 'rel='nofollow'>#{submission.pretty_format}</a>"
-      latest = ontology.explore.latest_submission({:include_status => 'ready'})
+      latest = ontology.explore.latest_submission({ :include_status => 'ready' })
       if latest && latest.submissionId == submission.submissionId
         link += " | <a href='#{ontology.id}/download?apikey=#{get_apikey}&download_format=csv' rel='nofollow'>CSV</a>"
         if !latest.hasOntologyLanguage.eql?("UMLS")
@@ -279,7 +279,7 @@ module OntologiesHelper
       end
       unless submission.diffFilePath.nil?
         uri = submission.id + "/download_diff?apikey=#{get_apikey}"
-        link = link + " | <a href='#{uri} 'rel='nofollow'>Diff</a>"
+        link = link + " | <a href='#{uri} 'rel='nofollow'>DIFF</a>"
       end
     end
     return link
@@ -296,13 +296,13 @@ module OntologiesHelper
   end
 
   # Creates a link based on the status of an ontology submission
-  def status_link(submission, sub_ontology=nil, latest=false, target="")
+  def status_link(submission, sub_ontology = nil, latest = false, target = "")
     version_text = submission.version.nil? || submission.version.length == 0 ? "unknown" : submission.version
     status_text = " <span class='ontology_submission_status'>" + submission_status2string(submission) + "</span>"
     if sub_ontology.nil?
       sub_ontology = submission.explore.ontology
     end
-    if sub_ontology.summaryOnly || latest==false
+    if sub_ontology.summaryOnly || latest == false
       version_link = version_text
     else
       version_link = "<a href='/ontologies/#{sub_ontology.acronym}?p=classes' #{target.empty? ? "" : "target='#{target}'"}>#{version_text}</a>"
@@ -318,12 +318,12 @@ module OntologiesHelper
     # Strip the URI prefix from the status codes (works even if they are not URIs)
     # The order of the codes must be assumed to be random, it is not an entirely
     # predictable sequence of ontology processing stages.
-    codes = sub.submissionStatus.map {|s| s.split('/').last }
-    errors = codes.select {|c| c.start_with? 'ERROR'}.map {|c| c.gsub("_", " ").split(/(\W)/).map(&:capitalize).join}.compact
+    codes = sub.submissionStatus.map { |s| s.split('/').last }
+    errors = codes.select { |c| c.start_with? 'ERROR' }.map { |c| c.gsub("_", " ").split(/(\W)/).map(&:capitalize).join }.compact
     status = []
     status.push('Parsed') if (codes.include? 'RDF') && (codes.include? 'RDF_LABELS')
     # The order of this array imposes an oder on the UI status code string
-    status_list = [ "INDEXED", "METRICS", "ANNOTATOR", "ARCHIVED" ]
+    status_list = ["INDEXED", "METRICS", "ANNOTATOR", "ARCHIVED"]
     status_list.insert(0, 'UPLOADED') unless status.include?('Parsed')
     status_list.each do |c|
       status.push(c.capitalize) if codes.include? c
@@ -335,9 +335,9 @@ module OntologiesHelper
 
   # Link for private/public/licensed ontologies
   def visibility_link(ontology)
-    ont_url = "/ontologies/#{ontology.acronym}"  # 'ontology' is NOT a submission here
-    page_name = 'summary'  # default ontology page view for visibility link
-    link_name = 'Public'   # default ontology visibility
+    ont_url = "/ontologies/#{ontology.acronym}" # 'ontology' is NOT a submission here
+    page_name = 'summary' # default ontology page view for visibility link
+    link_name = 'Public' # default ontology visibility
     if ontology.summaryOnly
       link_name = 'Summary Only'
     elsif ontology.private?
@@ -352,11 +352,11 @@ module OntologiesHelper
     ontology ||= @ontology
     return nil unless @analytics && @analytics[ontology.acronym.to_sym]
     return @visits_data if @visits_data
-    visits_data = {visits: [], labels: []}
-    years = @analytics[ontology.acronym.to_sym].to_h.keys.map {|e| e.to_s.to_i}.select {|e| e > 0}.sort
+    visits_data = { visits: [], labels: [] }
+    years = @analytics[ontology.acronym.to_sym].to_h.keys.map { |e| e.to_s.to_i }.select { |e| e > 0 }.sort
     now = Time.now
     years.each do |year|
-      months = @analytics[ontology.acronym.to_sym].to_h[year.to_s.to_sym].to_h.keys.map {|e| e.to_s.to_i}.select {|e| e > 0}.sort
+      months = @analytics[ontology.acronym.to_sym].to_h[year.to_s.to_sym].to_h.keys.map { |e| e.to_s.to_i }.select { |e| e > 0 }.sort
       months.each do |month|
         next if now.year == year && now.month <= month || (year == 2013 && month < 10) # we don't have good data going back past Oct 2013
         visits_data[:visits] << @analytics[ontology.acronym.to_sym].to_h[year.to_s.to_sym][month.to_s.to_sym]
@@ -366,10 +366,43 @@ module OntologiesHelper
     @visits_data = visits_data
   end
 
-
-
   def acronyms(ontologies)
     ontologies.present? ? ontologies.map { |ont| ont.acronym } : []
   end
 
+  def current_section
+    (params[:p]) ? params[:p] : 'summary'
+  end
+
+  def selected_section?(section_title)
+    current_section.eql?(section_title)
+  end
+
+  def lazy_load_section(section_title, &block)
+    render layout: 'ontologies/lazy_load_content',
+           locals: { current_section: current_section, section_title: section_title },
+           &block
+  end
+
+  def visits_chart_dataset(visits_data)
+    [{
+       label: 'Visits',
+       data: visits_data,
+       backgroundColor: 'rgba(151, 187, 205, 0.2)',
+       borderColor: 'rgba(151, 187, 205, 1)',
+       pointBorderColor: 'rgba(151, 187, 205, 1)',
+       pointBackgroundColor: 'rgba(151, 187, 205, 1)',
+     }].to_json
+  end
+
+  def sections_to_show
+    sections = ['summary']
+
+    unless @ontology.summaryOnly || @submission_latest.nil?
+      sections += %w[classes properties notes mappings]
+      sections += %w[instances] if @submission_latest.hasOntologyLanguage != 'SKOS'
+    end
+    sections
+  end
 end
+
