@@ -1,7 +1,7 @@
 class SubmissionsController < ApplicationController
 
   layout :determine_layout
-  before_action :authorize_and_redirect, :only=>[:edit,:update,:create,:new]
+  before_action :authorize_and_redirect, :only => [:edit, :update, :create, :new]
   before_action :submission_metadata, only: [:create, :edit, :new, :update]
 
   # When getting "Add submission" form to display
@@ -101,7 +101,7 @@ class SubmissionsController < ApplicationController
 
     @submission.update_from_params(submission_params)
     # Update summaryOnly on ontology object
-    @ontology.summaryOnly = @submission.isRemote.eql?("3")
+    @ontology.summaryOnly = @submission.isRemote.eql?('3')
     @ontology.update
     error_response = @submission.update(cache_refresh_all: false)
 
@@ -131,25 +131,31 @@ class SubmissionsController < ApplicationController
       :isRemote,
       :pullLocation,
       :filePath,
-      { contact:[:name, :email] },
+      { contact: [:name, :email] },
       :homepage,
       :documentation,
       :publication
     ]
-    
+
     @metadata.each do |m|
-      
+
       m_attr = m["attribute"].to_sym
-      
+
       attributes << if m["enforce"].include?("list")
-        { m_attr => [] }
+                      { m_attr => [] }
+                    else
+                      m_attr
+                    end
+    end
+
+    p = params.require(:submission).permit(attributes.uniq)
+    p.to_h.transform_values do |v|
+      if v.is_a? Array
+        v.reject(&:empty?)
       else
-        m_attr
+        v
       end
     end
-    
-    p = params.require(:submission).permit(attributes.uniq)
-    p.to_h
   end
 
 end
