@@ -14,16 +14,25 @@ Rails.application.routes.draw do
 
   resources :reviews
 
-  get '/mappings/new_external' => 'mappings#new_external'
+  get '/mappings/loader' , to: 'mappings#loader'
+  post '/mappings/loader', to: 'mappings#loader_process'
+  get 'mappings/count/:id', to: 'mappings#count', constraints: { id: /.+/ }
+  get 'mappings/show_mappings', to: 'mappings#show_mappings'
+  get 'mappings/new', to: 'mappings#new'
+  get 'mappings/:id', to: 'mappings#show', constraints: { id: /.+/ }
+  post 'mappings/:id', to: 'mappings#update', constraints: { id: /.+/ }
+  delete 'mappings/:id', to: 'mappings#destroy', constraints: { id: /.+/ }
   resources :mappings
 
   resources :margin_notes
 
   resources :concepts
 
+  get 'ontologies/:ontology_id/concepts', to: 'concepts#show_concept'
   resources :ontologies do
     resources :submissions
     get 'instances/:instance_id', to: 'instances#show', constraints: { instance_id: /[^\/?]+/ }
+    get 'schemes/show_scheme', to: 'schemes#show'
   end
 
   resources :login
@@ -63,21 +72,17 @@ Rails.application.routes.draw do
   match "/404", to: "errors#not_found", via: :all
   match "/500", to: "errors#internal_server_error", via: :all
 
-  # Analytics endpoint
-  get '/analytics' => 'analytics#track'
-
+  # Analytics
+  match 'analytics', to: 'analytics#track', via: [:post] 
+  
   # Robots.txt
   get '/robots.txt' => 'robots#index'
 
   # Ontologies
   get '/ontologies/view/edit/:id' => 'ontologies#edit_view', :constraints => { id: /[^\/?]+/ }
   get '/ontologies/view/new/:id' => 'ontologies#new_view'
-  get '/ontologies/virtual/:ontology
-
-
-
-
-' => 'ontologies#virtual', :as => :ontology_virtual
+  
+  get '/ontologies/virtual/:ontology' => 'ontologies#virtual', :as => :ontology_virtual
   get '/ontologies/success/:id' => 'ontologies#submit_success'
   match '/ontologies/:acronym' => 'ontologies#update', via: [:get, :post]
   match '/ontologies/:acronym/submissions/:id' => 'submissions#update', via: [:get, :post]
@@ -108,6 +113,7 @@ Rails.application.routes.draw do
   get '/ajax/classes/definition' => 'concepts#show_definition'
   get '/ajax/classes/treeview' => 'concepts#show_tree'
   get '/ajax/properties/tree' => 'concepts#property_tree'
+  get 'ajax/:ontology_id/schemes/:scheme_id/show_label', to: "schemes#show_label", constraints: { scheme_id: /[^\/?]+/ }
   get '/ajax/biomixer' => 'concepts#biomixer'
   get '/ajax/fair_score/html' => 'fair_score#details_html'
   get '/ajax/fair_score/json' => 'fair_score#details_json'
